% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{batch_summary}
\alias{batch_summary}
\title{BUSCO summary output for batch mode}
\format{
A 2-column data frame with the following variables:
\describe{
\item{Class}{Factor of BUSCO classes}
\item{Frequency}{Numeric with the percentage of BUSCOs in each class.}
\item{Lineage}{Character with the lineage dataset used.}
\item{File}{Character with the name of the FASTA file used.}
}
}
\usage{
data(batch_summary)
}
\description{
This object was created with the function \code{read_busco()} using
a batch run of BUSCO on the genomes of Herbaspirillum seropedicae SmR1
and Herbaspirillum rubrisubalbicans M1.
}
\examples{
data(batch_summary)
}
\keyword{datasets}
