% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{busco_is_installed}
\alias{busco_is_installed}
\title{Check if BUSCO is installed}
\usage{
busco_is_installed()
}
\value{
Logical indicating whether BUSCO is installed or not.
}
\description{
Check if BUSCO is installed
}
\examples{
busco_is_installed()
}
