% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogeqc-package.R
\docType{package}
\name{cogeqc-package}
\alias{cogeqc}
\alias{cogeqc-package}
\title{cogeqc: Systematic quality checks on comparative genomics analyses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

cogeqc aims to facilitate systematic quality checks on standard comparative genomics analyses to help researchers detect issues and select the most suitable parameters for each data set. cogeqc can be used to asses: i. genome assembly and annotation quality with BUSCOs and comparisons of statistics with publicly available genomes on the NCBI; ii. orthogroup inference using a protein domain-based approach and; iii. synteny detection using synteny network properties. There are also data visualization functions to explore QC summary statistics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/almeidasilvaf/cogeqc}
  \item Report bugs at \url{https://support.bioconductor.org/t/cogeqc}
}

}
\author{
\strong{Maintainer}: Fabrício Almeida-Silva \email{fabricio_almeidasilva@hotmail.com} (\href{https://orcid.org/0000-0002-5314-2964}{ORCID})

Authors:
\itemize{
  \item Yves Van de Peer \email{yves.vandepeer@psb.vib-ugent.be} (\href{https://orcid.org/0000-0003-4327-3730}{ORCID})
}

}
\keyword{internal}
