% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{fit_sft}
\alias{fit_sft}
\title{Goodness of fit test for the scale-free topology model}
\usage{
fit_sft(edges)
}
\arguments{
\item{edges}{A 2-column data frame with network edges represented in each.
Columns 1 and 2 represent nodes 1 and 2 of each edge.}
}
\value{
A numeric scalar with the R squared for the scale-free topology fit.
}
\description{
Goodness of fit test for the scale-free topology model
}
\examples{
data(synnet)
edges <- synnet
fit_sft(edges)
}
