% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{get_og_overlap}
\alias{get_og_overlap}
\title{Get number of shared orthogroups between species pairs}
\usage{
get_og_overlap(orthogroups)
}
\arguments{
\item{orthogroups}{A 3-column data frame with columns \strong{Orthogroup},
\strong{Species}, and \strong{Gene}. This data frame can be created from
the 'Orthogroups.tsv' file generated by OrthoFinder with the function
\code{read_orthogroups()}.}
}
\value{
A data frame with the following variables:
\describe{
\item{sp1}{Character, name of species 1.}
\item{sp2}{Character, name of species 2.}
\item{og_overlap}{Numeric, number of shared orthogroups.}
}
}
\description{
Get number of shared orthogroups between species pairs
}
\examples{
path <- system.file("extdata", "Orthogroups.tsv.gz", package = "cogeqc")
orthogroups <- read_orthogroups(path)
ov_df <- get_og_overlap(orthogroups)
}
