% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{interpro_bol}
\alias{interpro_bol}
\title{Intepro annotation for Brassica oleraceae's genes}
\format{
A 2-column data frame:
\describe{
\item{Gene}{Character of gene IDs.}
\item{Annotation}{Character of Interpro domains.}
}
}
\usage{
data(interpro_bol)
}
\description{
The annotation data were retrieved from PLAZA Dicots 5.0.
}
\examples{
data(interpro_bol)
}
\references{
Van Bel, M., Silvestri, F., Weitz, E. M., Kreft, L., Botzki, A.,
Coppens, F., & Vandepoele, K. (2021). PLAZA 5.0: extending the scope
and power of comparative and functional genomics in plants.
Nucleic acids research.
}
\keyword{datasets}
