% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_assembly.R
\name{list_busco_datasets}
\alias{list_busco_datasets}
\title{List BUSCO data sets}
\usage{
list_busco_datasets()
}
\value{
A hierarchically organized list of available data sets as returned
by \code{busco --list-datasets}.
}
\description{
List BUSCO data sets
}
\examples{
if(busco_is_installed()) {
    list_busco_datasets()
}
}
