% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{og}
\alias{og}
\title{Orthogroups between Arabidopsis thaliana and Brassica oleraceae}
\format{
A 3-column data frame with the following variables:
\describe{
\item{Orthogroup}{Orthogroup ID.}
\item{Species}{Abbreviation for species' name.}
\item{Gene}{Gene ID}
}
}
\usage{
data(og)
}
\description{
Data obtained from PLAZA Dicots 5.0.
}
\examples{
data(og)
}
\references{
Van Bel, M., Silvestri, F., Weitz, E. M., Kreft, L., Botzki, A.,
Coppens, F., & Vandepoele, K. (2021). PLAZA 5.0: extending the scope
and power of comparative and functional genomics in plants.
Nucleic acids research.
}
\keyword{datasets}
