% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{og_overlap_model}
\alias{og_overlap_model}
\title{Number of shared orthogroups between model organisms}
\format{
A 3-column data frame with the following variables:
\describe{
\item{sp1}{Character, name of species 1.}
\item{sp2}{Character, name of species 2.}
\item{og_overlap}{Numeric, number of shared orthogroups.}
}
}
\usage{
data(og_overlap_model)
}
\description{
Orthogroups from a set of model organisms obtained from
OrthoFinder2' example output, which was then passed
to \code{get_og_overlap()} to get shared orthogroups.
}
\examples{
data(og_overlap_model)
}
\references{
Emms, D. M., & Kelly, S. (2019). OrthoFinder: phylogenetic orthology
inference for comparative genomics. Genome biology, 20(1), 1-14.
}
\keyword{datasets}
