% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_duplications}
\alias{plot_duplications}
\title{Plot species-specific duplications}
\usage{
plot_duplications(stats_list = NULL)
}
\arguments{
\item{stats_list}{A list of data frames with Orthofinder summary stats
as returned by the function \code{read_orthofinder_stats}.}
}
\value{
A ggplot object with a barplot of number of species-specific
duplications.
}
\description{
Plot species-specific duplications
}
\examples{
dir <- system.file("extdata", package = "cogeqc")
stats_list <- read_orthofinder_stats(dir)
plot_duplications(stats_list)
}
