% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_genes_in_ogs}
\alias{plot_genes_in_ogs}
\title{Plot percentage of genes in orthogroups for each species}
\usage{
plot_genes_in_ogs(stats_list = NULL)
}
\arguments{
\item{stats_list}{A list of data frames with Orthofinder summary stats
as returned by the function \code{read_orthofinder_stats}.}
}
\value{
A ggplot object with a barplot of percentages of genes in
orthogroups for each species.
}
\description{
Plot percentage of genes in orthogroups for each species
}
\examples{
dir <- system.file("extdata", package = "cogeqc")
stats_list <- read_orthofinder_stats(dir)
plot_genes_in_ogs(stats_list)
}
