% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_og_overlap}
\alias{plot_og_overlap}
\title{Plot pairwise orthogroup overlap between species}
\usage{
plot_og_overlap(overlap_df, add_numbers = FALSE, abbreviate_names = TRUE)
}
\arguments{
\item{overlap_df}{A data frame with the number of shared orthogroups
between species pairs as returned by \code{get_og_overlap()}.}

\item{add_numbers}{Logical indicating whether to add numbers to each entry
of the heatmap. Default: FALSE.}

\item{abbreviate_names}{Logical indicating whether to abbreviate species
names. Default: TRUE.}
}
\value{
A ggplot object with a heatmap displaying the number of shared
orthogroups between species pairs.
}
\description{
Plot pairwise orthogroup overlap between species
}
\examples{
data(og_overlap_model)
plot_og_overlap(og_overlap_model, add_numbers = TRUE)
}
