% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_species_specific_ogs}
\alias{plot_species_specific_ogs}
\title{Plot number of species-specific orthogroups}
\usage{
plot_species_specific_ogs(stats_table = NULL)
}
\arguments{
\item{stats_table}{A 3-column data frame with Orthofinder summary stats
as returned by the function \code{read_orthofinder_stats}.}
}
\value{
A ggplot object with a barplot of number of species-specific
orthogroups for each species.
}
\description{
Plot number of species-specific orthogroups
}
\examples{
file <- system.file("extdata", "Statistics_PerSpecies.tsv",
                    package = "cogeqc")
stats_table <- read_orthofinder_stats(file)
plot_genes_in_ogs(stats_table)
}
