% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_orthogroups}
\alias{read_orthogroups}
\title{Read and parse orthogroups file created by OrthoFinder}
\usage{
read_orthogroups(orthogroups_path = NULL)
}
\arguments{
\item{orthogroups_path}{Path to Orthogroups/Orthogroups.tsv file generated
by OrthoFinder.}
}
\value{
A 3-column data frame with orthogroups, species IDs and
gene IDs, respectively.
}
\description{
This function converts the orthogroups file named \strong{Orthogroups.tsv} to
a parsed data frame.
}
\examples{
path <- system.file("extdata", "Orthogroups.tsv", package = "cogeqc")
og <- read_orthogroups(path)
}
\author{
Fabricio Almeida-Silva
}
