% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synnet}
\alias{synnet}
\title{Synteny network for Brassica oleraceae, B. napus, and B. rapa}
\format{
A 2-column data frame with the variables \strong{anchor1}
and \strong{anchor2}, containing names of loci in anchor 1 and anchor 2,
respectively.
}
\usage{
data(synnet)
}
\description{
Synteny network for Brassica oleraceae, B. napus, and B. rapa
}
\examples{
data(synnet)
}
\references{
Zhao, T., & Schranz, M. E. (2019). Network-based microsynteny
analysis identifies major differences and genomic outliers
in mammalian and angiosperm genomes. Proceedings of the National Academy
of Sciences, 116(6), 2165-2174.
}
\keyword{datasets}
