% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tree}
\alias{tree}
\title{Species tree for model species}
\format{
An object of class "phylo" as returned by \code{treeio::read.tree()}.
}
\usage{
data(tree)
}
\description{
The data used to create this object was retrieved from Orthofinder's
example output for model species, available in
https://bioinformatics.plants.ox.ac.uk/davidemms/public_data/.
}
\examples{
data(tree)
}
\references{
Emms, D. M., & Kelly, S. (2019). OrthoFinder: phylogenetic orthology
inference for comparative genomics. Genome biology, 20(1), 1-14.
}
\keyword{datasets}
