% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synteny.R
\name{assess_synnet}
\alias{assess_synnet}
\title{Assess synteny network based on clustering coefficient and node frequency}
\usage{
assess_synnet(synnet = NULL, cc_type = "average")
}
\arguments{
\item{synnet}{Edgelist for the synteny network in a 2-column data frame,
with variables \strong{anchor1} and \strong{anchor2} representing
names of loci in anchor 1 and anchor 2, respectively.}

\item{cc_type}{Type of clustering coefficient to be calculated.
One of 'global' or 'average'. Default: 'average'.}
}
\value{
A data frame with the following variables:
\describe{
\item{CC}{Numeric representing clustering coefficient.}
\item{Node_number}{Numeric representing number of nodes in the network.}
\item{Score}{Numeric representing network score, which is
the product of 'CC' and 'Node_number'.}
}
}
\description{
Assess synteny network based on clustering coefficient and node frequency
}
\examples{
data(synnet)
assess_synnet(synnet)
}
