% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homology_detection.R
\name{calculate_H}
\alias{calculate_H}
\title{Calculate homogeneity scores for orthogroups}
\usage{
calculate_H(orthogroup_df = NULL)
}
\arguments{
\item{orthogroup_df}{Data frame with orthogroups and their associated genes
and annotation. The columns \strong{Gene}, \strong{Orthogroup}, and
\strong{Annotation} are mandatory, and they must represent Gene ID,
Orthogroup ID, and Annotation ID (e.g., Interpro/PFAM), respectively.}
}
\value{
A 2-column data frame with the variables \strong{Orthogroup}
and \strong{mean_H}, corresponding to orthogroup ID and mean homogeneity,
respectively.
}
\description{
Calculate homogeneity scores for orthogroups
}
\examples{
data(og)
data(interpro_ath)
orthogroup_df <- merge(og[og$Species == "Ath", ], interpro_ath)
H <- calculate_H(orthogroup_df)
}
