% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_assembly.R
\name{list_busco_datasets}
\alias{list_busco_datasets}
\title{List BUSCO data sets}
\usage{
list_busco_datasets(envname = NULL, miniconda_path = NULL)
}
\arguments{
\item{envname}{Name of the Conda environment with external dependencies
to be included in the temporary R environment.}

\item{miniconda_path}{Path to miniconda. Only valid if envname is specified.}
}
\value{
A hierarchically organized list of available data sets as returned
by \code{busco --list-datasets}.
}
\description{
List BUSCO data sets
}
\examples{
if(busco_is_installed()) {
    list_busco_datasets()
}
}
