% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_og_overlap}
\alias{plot_og_overlap}
\title{Plot pairwise orthogroup overlap between species}
\usage{
plot_og_overlap(stats_list = NULL, clust = TRUE)
}
\arguments{
\item{stats_list}{A list of data frames with Orthofinder summary stats
as returned by the function \code{read_orthofinder_stats}.}

\item{clust}{Logical indicating whether to clust data based on overlap.
Default: TRUE}
}
\value{
A ggplot object with a heatmap.
}
\description{
Plot pairwise orthogroup overlap between species
}
\examples{
dir <- system.file("extdata", package = "cogeqc")
stats_list <- read_orthofinder_stats(dir)
plot_og_overlap(stats_list)
}
