% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_species_specific_ogs}
\alias{plot_species_specific_ogs}
\title{Plot number of species-specific orthogroups}
\usage{
plot_species_specific_ogs(stats_list = NULL)
}
\arguments{
\item{stats_list}{A list of data frames with Orthofinder summary stats
as returned by the function \code{read_orthofinder_stats}.}
}
\value{
A ggplot object with a barplot of number of species-specific
orthogroups for each species.
}
\description{
Plot number of species-specific orthogroups
}
\examples{
dir <- system.file("extdata", package = "cogeqc")
stats_list <- read_orthofinder_stats(dir)
plot_species_specific_ogs(stats_list)
}
