% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import.R
\name{read_orthofinder_stats}
\alias{read_orthofinder_stats}
\title{Read and parse per-species summary statistics in a data frame}
\usage{
read_orthofinder_stats(stats_path = NULL)
}
\arguments{
\item{stats_path}{Path to file containing Orthofinder's per-species
stats. In your Orthofinder results directory, this file is located
at \strong{Comparative_Genomics_Statistics/Statistics_PerSpecies.tsv}.}
}
\value{
A data frame in long format with the following variables:
\describe{
\item{Species}{Factor of species names.}
\item{N_genes}{Numeric of number of genes.}
\item{N_genes_in_OGs}{Numeric of number of genes in orthogroups.}
\item{N_unassigned}{Numeric of number of unassigned genes (not in OGs).}
\item{Perc_genes_in_OGs}{Numeric of percentage of genes in orthogroups.}
\item{Perc_unassigned}{Numeric of percentage of unassigned genes.}
\item{N_ss_OGs}{Numeric of number of species-specific orthogroups.}
\item{N_genes_in_ss_OGs}{Numeric of number of genes in species-specific
orthogroups.}
\item{Perc_genes_in_ss_OGs}{Numeric of percentage of genes in
species-specific orthogroups.}
}
}
\description{
Read and parse per-species summary statistics in a data frame
}
\examples{
stats_path <- system.file("extdata", "Statistics_PerSpecies.tsv",
                          package = "cogeqc")
ortho_stats <- read_orthofinder_stats(stats_path)
}
